requirements.txt
discord.py>=2.4.0
PyYAML>=6.0.1
APScheduler>=3.10.4
python-dotenv>=1.0.1
google-genai>=1.30.0
pytest>=8.3.0
certifi>=2026.1.4
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
streamlit>=1.40.0
httpx>=0.28.0
feedparser>=6.0.11
PyNaCl>=1.5.0
yt-dlp>=2025.1.26
